Centralize bootstrap_trace
